# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Системы для управления банковскими картами",
    "contact": {
      "name": "Вурц Геннадий",
      "email": "vgd2001@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "Админ: Пользователи"
        ],
        "summary": "Поиск пользователей",
        "description": "Поддерживается фильтрация и пагинация",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Имя пользователя, поддерживается поиск по части имени",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Иван"
          },
          {
            "name": "email",
            "in": "query",
            "description": "Email пользователя, поддерживается поиск по части email",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "@mail"
          },
          {
            "name": "from",
            "in": "query",
            "description": "Индекс первого элемента, с которого нужно вернуть",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Количество возвращаемых элементов",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователи найдены",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "examples": [
                    "{\"error\":\"from должно быть не меньше 0.\"}"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Админ: Пользователи"
        ],
        "summary": "Создание нового пользователя",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Пользователь создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "examples": [
                    "{\"error\": \"Поле password размер должен находиться в диапазоне от 8 до 255.\"}\n"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{userId}": {
      "get": {
        "tags": [
          "Админ: Пользователи"
        ],
        "summary": "Получение пользователя по его Id",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Id пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFullDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "User with id 42 not found"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse",
                  "example": {
                    "error": "userId должно быть больше 0."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Сведения об ошибке",
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки",
            "example": "error message"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "description": "Информация о пользователе",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "id пользователя",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "имя пользователя",
            "example": "Иван Иванов"
          },
          "email": {
            "type": "string",
            "description": "email пользователя",
            "example": "ivan@mail.ru"
          },
          "role": {
            "type": "string",
            "description": "роль пользователя",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "example": "USER"
          }
        }
      },
      "UserCreateRequest": {
        "type": "object",
        "description": "Данные для создания нового пользователя",
        "properties": {
          "name": {
            "type": "string",
            "description": "имя пользователя",
            "example": "Иван Иванов",
            "maxLength": 255,
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "email пользователя",
            "example": "ivan@mail.ru",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "description": "роль пользователя",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "example": "USER"
          },
          "password": {
            "type": "string",
            "description": "пароль пользователя",
            "example": "qwerty123",
            "maxLength": 255,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "name",
          "password",
          "role"
        ]
      },
...
      "CardShortDto": {
        "type": "object",
        "description": "Информация о банковской карте, не включающая информацию о владельце",
        "properties": {
          "number": {
            "type": "string",
            "description": "номер карты",
            "example": "**** **** **** 1234"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "дата создания карты",
            "example": "2025-09-28T12:00"
          },
          "expiration": {
            "type": "string",
            "format": "date",
            "description": "дата окончания срока действия карты",
            "example": "2026-09-28"
          },
          "status": {
            "type": "string",
            "description": "статус карты",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED"
            ],
            "example": "ACTIVE"
          },
          "balance": {
            "type": "number",
            "description": "баланс карты",
            "example": 1234.56
          }
        }
      },
      "UserFullDto": {
        "type": "object",
        "description": "Информация о пользователе, включающая информацию о его картах",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "id пользователя",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "имя пользователя",
            "example": "Иван Иванов"
          },
          "email": {
            "type": "string",
            "description": "email пользователя",
            "example": "ivan@mail.ru"
          },
          "role": {
            "type": "string",
            "description": "роль пользователя",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "example": "USER"
          },
          "cards": {
            "type": "array",
            "description": "список карт пользователя",
            "items": {
              "$ref": "#/components/schemas/CardShortDto"
            }
          }
        }
      },
...
    },
...
  }
}